update and fixing bugs

+++ Implementation/UseCases/Commands/Products/EfUpdatProduct.cs
@@ -32,9 +32,9 @@ namespace Implementation.UseCases.Commands.Products
             validation.ValidateAndThrow(data);
             Product product = Context.Products.Find(data.Id);
             if (product == null) { throw new EntityNotFoundException(nameof(Product), data.Id); }
-            var colors = JsonConvert.DeserializeObject<List<int>>(data.ProductColors);
+            
             // Deserializacija JSON stringa u listu ProductSpecificationDTO objekata
-            var Specifications = JsonConvert.DeserializeObject<List<ProductSpecificationDTO>>(data.ProductSpecifications);
+            
             string profile_path = PictureUpload.Upload(data.Image, "products");
             product.Name = data.ProductName;
             product.Description= data.ProductDescription;
@@ -43,29 +43,38 @@ namespace Implementation.UseCases.Commands.Products
             product.Gender_id= data.GenderId;
             Product_Price price=Context.Product_Prices.FirstOrDefault(y=>y.Product_id==data.Id);
             price.Price = data.ProductPrice;
-            var color = Context.Product_Colors.Where(y => y.Product_id == data.Id).ToList();
-            Context.Product_Colors.RemoveRange(color);
-            foreach (var item in colors)
+            if (!string.IsNullOrEmpty(data.ProductColors))
             {
-                var newColor = new Product_Color
+                var colors = JsonConvert.DeserializeObject<List<int>>(data.ProductColors);
+                var color = Context.Product_Colors.Where(y => y.Product_id == data.Id).ToList();
+                Context.Product_Colors.RemoveRange(color);
+                foreach (var item in colors)
                 {
-                    Color_id = item,
-                    Product_id = data.Id,
+                    var newColor = new Product_Color
+                    {
+                        Color_id = item,
+                        Product_id = data.Id,
+                    };
+                    Context.Product_Colors.Add(newColor);
                 };
-                Context.Product_Colors.Add(newColor);
-            };
-            var old_specifications = Context.Product_Specifications.Where(y => y.Product_id == data.Id).ToList();
-            Context.Product_Specifications.RemoveRange(old_specifications);
-            foreach(var item2 in Specifications)
-            {
-                var newSpecification = new Product_Specification
-                {
-                    Product_id = data.Id,
-                    Specification_id = item2.SpecificationId,
-                    Value = item2.SpecificationValue
-                };
-                Context.Product_Specifications.Add(newSpecification);
             }
+            if (!string.IsNullOrEmpty(data.ProductSpecifications))
+            {
+                var Specifications = JsonConvert.DeserializeObject<List<ProductSpecificationDTO>>(data.ProductSpecifications);
+                var old_specifications = Context.Product_Specifications.Where(y => y.Product_id == data.Id).ToList();
+                Context.Product_Specifications.RemoveRange(old_specifications);
+                foreach (var item2 in Specifications)
+                {
+                    var newSpecification = new Product_Specification
+                    {
+                        Product_id = data.Id,
+                        Specification_id = item2.SpecificationId,
+                        Value = item2.SpecificationValue
+                    };
+                    Context.Product_Specifications.Add(newSpecification);
+                }
+            }
+            
             Context.SaveChanges();
